requirements.txt
appnope==0.1.3
asgiref==3.5.0
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
cffi==1.15.0
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
debugtools==0.2.0
decorator==5.1.1
devtools==0.8.0
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djlint==0.7.6
executing==0.8.3
fabric==2.7.0
gunicorn==20.1.0
importlib-metadata==4.11.3
invoke==1.7.0
ipython==8.2.0
jedi==0.18.1
Markdown==3.3.6
matplotlib-inline==0.1.3
paramiko==2.10.3
parso==0.8.3
pathlib2==2.3.7.post1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
PyNaCl==1.5.0
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomlkit==0.9.2
tqdm==4.64.0
traitlets==5.1.1
wcwidth==0.2.5
zipp==3.8.0
